Implemented menu module
